+++ test/e2e/tests/onboarding/onboarding.spec.ts
@@ -27,16 +27,6 @@ import {
 import { switchToNetworkFlow } from '../../page-objects/flows/network.flow';
 
 describe('MetaMask onboarding', function () {
-  const ganacheOptions2 = {
-    accounts: [
-      {
-        secretKey:
-          '0x53CB0AB5226EEBF4D872113D98332C1555DC304443BEE1CF759D15798D3C55A9',
-        balance: convertToHexValue(10000000000000000000),
-      },
-    ],
-  };
-
   it("Creates a new wallet, sets up a secure password, and doesn't complete the onboarding process and refreshes the page", async function () {
     await withFixtures(
       {
@@ -225,20 +215,23 @@ describe('MetaMask onboarding', function () {
         fixtures: new FixtureBuilder({ onboarding: true }).build(),
         localNodeOptions: [
           {
-            type: 'ganache',
+            type: 'anvil',
           },
           {
-            type: 'ganache',
+            type: 'anvil',
             options: {
               port,
               chainId,
-              ...ganacheOptions2,
             },
           },
         ],
         title: this.test?.fullTitle(),
       },
       async ({ driver, localNodes }) => {
+        await localNodes[1].setAccountBalance(
+          '0x0Cc5261AB8cE458dc977078A3623E2BaDD27afD3',
+          convertToHexValue(10000000000000000000),
+        );
         await importSRPOnboardingFlow({
           driver,
           seedPhrase: TEST_SEED_PHRASE,
@@ -270,9 +263,9 @@ describe('MetaMask onboarding', function () {
 
         // Check the correct balance for the custom network is displayed
         if (localNodes[1] && Array.isArray(localNodes)) {
-          await homePage.check_localNodeBalanceIsDisplayed(localNodes[1]);
+          await homePage.check_expectedBalanceIsDisplayed('10');
         } else {
-          throw new Error('Custom network Ganache server not available');
+          throw new Error('Custom network server not available');
         }
       },
     );