+++ ui/ducks/ramps/ramps.ts
@@ -4,7 +4,10 @@ import { getCurrentChainId } from '../../../shared/modules/selectors/networks';
 import { getUseExternalServices } from '../../selectors';
 import RampAPI from '../../helpers/ramps/rampApi/rampAPI';
 import { hexToDecimal } from '../../../shared/modules/conversion.utils';
-import { getMultichainIsBitcoin } from '../../selectors/multichain';
+import {
+  getMultichainIsBitcoin,
+  getMultichainIsSolana,
+} from '../../selectors/multichain';
 import { MultichainNetworks } from '../../../shared/constants/multichain/networks';
 import { defaultBuyableChains } from './constants';
 import { AggregatorNetwork } from './types';
@@ -68,6 +71,8 @@ const rampsSlice = createSlice({
 const { reducer } = rampsSlice;
 
 // Can be typed to RootState if/when the interface is defined
+
+// TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31973
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const getBuyableChains = (state: any) =>
   state.ramps?.buyableChains ?? defaultBuyableChains;
@@ -83,23 +88,49 @@ export const getIsBitcoinBuyable = createSelector(
       ),
 );
 
+export const getIsSolanaBuyable = createSelector(
+  [getBuyableChains],
+  (buyableChains) => {
+    return buyableChains
+      .filter(Boolean)
+      .some(
+        (network: AggregatorNetwork) =>
+          network.chainId === MultichainNetworks.SOLANA && network.active,
+      );
+  },
+);
+
 export const getIsNativeTokenBuyable = createSelector(
   [
     getCurrentChainId,
     getBuyableChains,
     getIsBitcoinBuyable,
     getMultichainIsBitcoin,
+    getIsSolanaBuyable,
+    getMultichainIsSolana,
   ],
-  (currentChainId, buyableChains, isBtcBuyable, isBtc) => {
+  (
+    currentChainId,
+    buyableChains,
+    isBtcBuyable,
+    isBtc,
+    isSolanaBuyable,
+    isSolana,
+  ) => {
     try {
+      if (isBtc) {
+        return isBtcBuyable;
+      }
+      if (isSolana) {
+        return isSolanaBuyable;
+      }
+
       return buyableChains
         .filter(Boolean)
-        .some((network: AggregatorNetwork) => {
-          if (isBtc) {
-            return isBtcBuyable;
-          }
-          return String(network.chainId) === hexToDecimal(currentChainId);
-        });
+        .some(
+          (network: AggregatorNetwork) =>
+            String(network.chainId) === hexToDecimal(currentChainId),
+        );
     } catch (e) {
       return false;
     }